Refactor the code of schedulers. * Split vineyard schedulers into two files to make the code clear. * Add NOTES.txt to helm chart.
