+++ charts/vineyard-operator/templates/NOTES.txt
@@ -1,0 +1,8 @@
+Thanks for installing {{ .Chart.Name | upper }}, release at namespace: {{ .Release.Namespace }}, name: {{ .Release.Name }}.
+
+To learn more about the release, try: 
+
+  $ helm status {{ .Release.Name }} -n {{ .Release.Namespace }}   # get status of running vineyard operator
+  $ helm get all {{ .Release.Name }} -n {{ .Release.Namespace }}  # get all deployment yaml of vineyard operator
+
+Just have fun with vineyard operator!
